+++ amd_openvx_extensions/amd_rpp/source/tensor/Normalize.cpp
@@ -178,6 +178,7 @@ static vx_status VX_CALLBACK processNormalize(vx_node node, const vx_reference *
     NormalizeLocalData *data = NULL;
     STATUS_ERROR_CHECK(vxQueryNode(node, VX_NODE_LOCAL_DATA_PTR, &data, sizeof(data)));
     refreshNormalize(node, parameters, num, data);
+#if RPP_AUDIO
     if (data->deviceType == AGO_TARGET_AFFINITY_GPU) {
 #if ENABLE_HIP
         rpp_status = rppt_normalize_gpu(data->pSrc, data->pSrcGenericDesc, data->pDst, data->pDstGenericDesc, data->axis_mask, data->pMean, data->pStddev, data->computeMeanAndStdDev, data->scale, data->shift, data->pSrcRoi, data->handle->rppHandle);
@@ -188,6 +189,7 @@ static vx_status VX_CALLBACK processNormalize(vx_node node, const vx_reference *
         rpp_status = rppt_normalize_host(data->pSrc, data->pSrcGenericDesc, data->pDst, data->pDstGenericDesc, data->axis_mask, data->pMean, data->pStddev, data->computeMeanAndStdDev, data->scale, data->shift, data->pSrcRoi, data->handle->rppHandle);
         return_status = (rpp_status == RPP_SUCCESS) ? VX_SUCCESS : VX_FAILURE;
     }
+#endif
     return return_status;
 }
 
@@ -242,12 +244,16 @@ static vx_status VX_CALLBACK uninitializeNormalize(vx_node node, const vx_refere
     STATUS_ERROR_CHECK(releaseRPPHandle(node, data->handle, data->deviceType));
     if (data->deviceType == AGO_TARGET_AFFINITY_GPU) {
 #if ENABLE_HIP
-        hipError_t err = hipHostFree(data->pMean);
-        if (err != hipSuccess)
-            std::cerr << "\n[ERR] hipFree failed  " << std::to_string(err) << "\n";
-        err = hipHostFree(data->pStddev);
-        if (err != hipSuccess)
-            std::cerr << "\n[ERR] hipFree failed  " << std::to_string(err) << "\n";
+        if (data->pMean) {
+            hipError_t err = hipHostFree(data->pMean);
+            if (err != hipSuccess)
+                std::cerr << "\n[ERR] hipHostFree failed  " << std::to_string(err) << "\n";
+        }
+        if (data->pStddev) {
+            hipError_t err = hipHostFree(data->pStddev);
+            if (err != hipSuccess)
+                std::cerr << "\n[ERR] hipHostFree failed  " << std::to_string(err) << "\n";
+        }
 #endif
     } else {
         if (data->pMean) delete[] data->pMean;