+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Documentation for MIVisionX is available at
 ### Optimizations
 
 * Readme
+* Setup - CentOS 7 support
 
 ### Changes
 
@@ -36,7 +37,7 @@ Documentation for MIVisionX is available at
   + Ubuntu - `20.04` / `22.04`
   + CentOS - `7`
   + RHEL - `8` / `9`
-  + SLES - `15-SP4`
+  + SLES - `15-SP5`
 * ROCm: rocm-core - `6.1.0.60100`
 * RPP - `1.5.0.60100`
 * miopen-hip - `3.1.0.60100`
@@ -44,7 +45,7 @@ Documentation for MIVisionX is available at
 * OpenCV - [4.6.0](https://github.com/opencv/opencv/releases/tag/4.6.0)
 * FFMPEG - [n4.4.2](https://github.com/FFmpeg/FFmpeg/releases/tag/n4.4.2)
 * Dependencies for all the above packages
-* MIVisionX Setup Script - `V3.2.0`
+* MIVisionX Setup Script - `V3.5.0`
 
 ### Known issues